Moved the heuristic to the ChessGame class so that it's updated for each movement/capture.
